+++ backend/test/routes/picker_options.test.js
@@ -12,4 +12,16 @@ testRoutes([
     capabilities: capability,
     type: routeTypes.General,
   },
+  {
+    request: { method: "GET", url: "/picker_options/project/1" },
+    modelFunction: model.findAllByProject,
+    capabilities: capability,
+    type: routeTypes.General,
+  },
+  {
+    request: { method: "GET", url: "/picker_options/contract/1" },
+    modelFunction: model.findAllByContract,
+    capabilities: capability,
+    type: routeTypes.General,
+  },
 ]);