+++ frontend/src/components/EditForm/index.tsx
@@ -24,7 +24,16 @@ export const EditForm = ({
           <Form>
             <FormLayout>
               {editFields.map(
-                ({ fieldName, fieldType, fieldLabel, width, tableName, pickerName, projectId }) => {
+                ({
+                  fieldName,
+                  fieldType,
+                  fieldLabel,
+                  width,
+                  tableName,
+                  pickerName,
+                  projectId,
+                  contractId,
+                }) => {
                   return (
                     <FormInput
                       setFieldValue={setFieldValue}
@@ -38,6 +47,7 @@ export const EditForm = ({
                       tableName={tableName}
                       pickerName={pickerName}
                       projectId={projectId}
+                      contractId={contractId}
                     />
                   );
                 }