+++ frontend/src/pages/Contracts/Contract/InvoiceProcessing/InvoiceDeliverables/fields.tsx
@@ -10,29 +10,49 @@ import { IEditFields } from "types";
  */
 export const readFields = (query: UseQueryResult<FormikValues>) => {
   return [
-    { width: "half", title: "Deliverables / Expense", value: query?.data?.resource_id },
-    //{ width: "half", title: "Type", value: query?.data?.hours },
-    { width: "half", title: "Amount", value: query?.data?.rate * query?.data?.rate },
+    {
+      width: "half",
+      title: "Deliverables / Expense",
+      value: query?.data?.contract_deliverable_id?.label,
+    },
+    { width: "half", title: "Fiscal Year", value: query?.data?.fiscal_year },
+    { width: "half", title: "Amount", value: query?.data?.rate },
+    { width: "half", title: "Amount Remaining", value: query?.data?.amount_remaining },
   ];
 };
 
 /**
  * The edit fields.
  *
+ * @param   {number}        contractId Id of contract to get deliverables for.
  * @returns {IEditFields[]}
  */
-export const editFields: IEditFields[] = [
-  {
-    width: "half",
-    fieldLabel: "Deliverables / Expense",
-    fieldName: "deliverable_id",
-    fieldType: "singleText",
-  },
-];
+export const editFields: (contractId: number) => IEditFields[] = (contractId) => {
+  return [
+    {
+      width: "half",
+      fieldLabel: "Deliverables / Expense",
+      fieldName: "contract_deliverable_id",
+      fieldType: "select",
+      pickerName: "contract_deliverable",
+      contractId: contractId,
+    },
+    { width: "half", fieldLabel: "Fiscal Year", fieldName: "fiscal_year", fieldType: "readonly" },
+    { width: "half", fieldLabel: "Amount", fieldName: "rate", fieldType: "number" },
+    {
+      width: "half",
+      fieldLabel: "Amount Remaining",
+      fieldName: "amount_remaining",
+      fieldType: "readonly",
+    },
+  ];
+};
 
 /**
  * Inital values for create form.
  */
 export const initialValues = {
-  deliverable_id: "",
+  contract_deliverable_id: "",
+  rate: 0,
+  unit_amount: 1,
 };